Add createImage, getImage, update getUser, finish refactor
