+++ server/functions/src/functions/getUser.ts
@@ -1,4 +1,4 @@
-import { CallableRequest, HttpsError } from "firebase-functions/https";
+import { CallableRequest } from "firebase-functions/https";
 import * as z from "zod";
 import * as util from "../util";
 import { logger } from "firebase-functions";
